weapon in safe
If the gun is taken from the weapon safe, a panel appears covering the whole screen with information about the gun.

As long as the panel is opened, the player is stopped from moving and timescale is set to 0, stopping all movements around the players enviroment.

Only the mouse cursor movement is enabled.

If player closes the panel, all movements are enabled and the door to go outside opens.