Objective when game starts
When game starts a panel appears covering the whole screen with the text that the player has 5 minutes to collect 4 keycards. On fail player dies.

As long as the panel is opened, the player is stopped from moving and timescale is set to 0, stopping all movements around the players enviroment.

Only the mouse cursor movement is enabled.

If player closes the panel, all movements are enabled.